added sign out and settings page
