+++ src/components/Layout.tsx
@@ -1,8 +1,21 @@
 import { Navbar } from './Navbar/Navbar';
 import styles from './Layout.module.css'
 import Head from "next/head";
+import { useRouter } from 'next/router';
+import { useEffect } from 'react';
+import { useAuth } from '../firebase/auth/useFirebaseAuth';
 
 export default function Layout({ children }) {
+  const { authUser, loading } = useAuth();
+  const router = useRouter();
+
+  // Listen for changes on loading and authUser, redirect if needed
+  useEffect(() => {
+    if (!loading && !authUser) {
+      router.push('/welcome')
+    }
+  });
+  
   return (
     <div className={styles.root}>
       <Head>
