+++ src/components/Navbar/Navbar.module.css
@@ -2,6 +2,7 @@
   display: flex; 
   align-items: center; 
   justify-content: flex-end; 
+  padding-right: 20px;
   width: 100%;
   height: 50px; 
   background-color: var(--prussian-blue); 
@@ -24,4 +25,11 @@
   color: var(--baby-pink); 
   border-bottom: 3px solid var(--baby-pink); 
   padding-bottom: 3px; 
+}
+
+.signout {
+  background-color: var(--prussian-blue); 
+  color: var(--cultured); 
+  width: 80px; 
+  border: none; 
 }
